requirements.txt
pyTelegramBotAPI==4.14.0
requests==2.31.0
APScheduler==3.10.4
python-dotenv==1.0.0
